Move events into Domain namespace
