+++ src/Sulu/Bundle/CustomUrlBundle/Domain/Event/CustomUrlRemovedEvent.php
@@ -9,56 +9,43 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\CustomUrlBundle\Event;
+namespace Sulu\Bundle\CustomUrlBundle\Domain\Event;
 
 use Sulu\Bundle\CustomUrlBundle\Admin\CustomUrlAdmin;
 use Sulu\Bundle\EventLogBundle\Event\DomainEvent;
-use Sulu\Component\CustomUrl\Document\CustomUrlDocument;
 
-class CustomUrlRouteRemovedEvent extends DomainEvent
+class CustomUrlRemovedEvent extends DomainEvent
 {
     /**
-     * @var CustomUrlDocument
+     * @var string
      */
-    private $customUrlDocument;
+    private $customUrlUuid;
+
+    /**
+     * @var string
+     */
+    private $customUrlTitle;
 
     /**
      * @var string
      */
     private $webspaceKey;
 
-    /**
-     * @var string
-     */
-    private $routeUuid;
-
     public function __construct(
-        CustomUrlDocument $customUrlDocument,
-        string $webspaceKey,
-        string $routeUuid
+        string $customUrlUuid,
+        string $customUrlTitle,
+        string $webspaceKey
     ) {
         parent::__construct();
 
-        $this->customUrlDocument = $customUrlDocument;
+        $this->customUrlUuid = $customUrlUuid;
+        $this->customUrlTitle = $customUrlTitle;
         $this->webspaceKey = $webspaceKey;
-        $this->routeUuid = $routeUuid;
-    }
-
-    public function getCustomUrlDocument(): CustomUrlDocument
-    {
-        return $this->customUrlDocument;
     }
 
     public function getEventType(): string
     {
-        return 'route_removed';
-    }
-
-    public function getEventPayload(): array
-    {
-        return [
-            'routeUuid' => $this->routeUuid,
-        ];
+        return 'removed';
     }
 
     public function getResourceKey(): string
@@ -68,7 +55,7 @@ class CustomUrlRouteRemovedEvent extends DomainEvent
 
     public function getResourceId(): string
     {
-        return (string) $this->customUrlDocument->getUuid();
+        return $this->customUrlUuid;
     }
 
     public function getResourceWebspaceKey(): ?string
@@ -78,7 +65,7 @@ class CustomUrlRouteRemovedEvent extends DomainEvent
 
     public function getResourceTitle(): ?string
     {
-        return $this->customUrlDocument->getTitle();
+        return $this->customUrlTitle;
     }
 
     public function getResourceSecurityContext(): ?string
