+++ src/Sulu/Bundle/CustomUrlBundle/Domain/Event/CustomUrlRouteRemovedEvent.php
@@ -9,13 +9,13 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\CustomUrlBundle\Event;
+namespace Sulu\Bundle\CustomUrlBundle\Domain\Event;
 
 use Sulu\Bundle\CustomUrlBundle\Admin\CustomUrlAdmin;
 use Sulu\Bundle\EventLogBundle\Event\DomainEvent;
 use Sulu\Component\CustomUrl\Document\CustomUrlDocument;
 
-class CustomUrlModifiedEvent extends DomainEvent
+class CustomUrlRouteRemovedEvent extends DomainEvent
 {
     /**
      * @var CustomUrlDocument
@@ -28,20 +28,20 @@ class CustomUrlModifiedEvent extends DomainEvent
     private $webspaceKey;
 
     /**
-     * @var array
+     * @var string
      */
-    private $payload;
+    private $routeUuid;
 
     public function __construct(
         CustomUrlDocument $customUrlDocument,
         string $webspaceKey,
-        array $payload
+        string $routeUuid
     ) {
         parent::__construct();
 
         $this->customUrlDocument = $customUrlDocument;
         $this->webspaceKey = $webspaceKey;
-        $this->payload = $payload;
+        $this->routeUuid = $routeUuid;
     }
 
     public function getCustomUrlDocument(): CustomUrlDocument
@@ -51,12 +51,14 @@ class CustomUrlModifiedEvent extends DomainEvent
 
     public function getEventType(): string
     {
-        return 'modified';
+        return 'route_removed';
     }
 
     public function getEventPayload(): array
     {
-        return $this->payload;
+        return [
+            'routeUuid' => $this->routeUuid,
+        ];
     }
 
     public function getResourceKey(): string
