+++ src/Sulu/Bundle/TagBundle/Domain/Event/TagCreatedEvent.php
@@ -9,12 +9,12 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\TagBundle\Event;
+namespace Sulu\Bundle\TagBundle\Domain\Event;
 
 use Sulu\Bundle\EventLogBundle\Event\DomainEvent;
 use Sulu\Bundle\TagBundle\Tag\TagInterface;
 
-class TagModifiedEvent extends DomainEvent
+class TagCreatedEvent extends DomainEvent
 {
     /**
      * @var TagInterface
@@ -43,7 +43,7 @@ class TagModifiedEvent extends DomainEvent
 
     public function getEventType(): string
     {
-        return 'modified';
+        return 'created';
     }
 
     public function getEventPayload(): array
