+++ src/Sulu/Bundle/TagBundle/Tag/TagManager.php
@@ -14,14 +14,14 @@ namespace Sulu\Bundle\TagBundle\Tag;
 use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Doctrine\Persistence\ObjectManager;
 use Sulu\Bundle\EventLogBundle\Collector\DomainEventCollectorInterface;
+use Sulu\Bundle\TagBundle\Domain\Event\TagCreatedEvent;
+use Sulu\Bundle\TagBundle\Domain\Event\TagMergedEvent;
+use Sulu\Bundle\TagBundle\Domain\Event\TagModifiedEvent;
+use Sulu\Bundle\TagBundle\Domain\Event\TagRemovedEvent;
 use Sulu\Bundle\TagBundle\Entity\TagRepository;
-use Sulu\Bundle\TagBundle\Event\TagCreatedEvent;
 use Sulu\Bundle\TagBundle\Event\TagDeleteEvent;
 use Sulu\Bundle\TagBundle\Event\TagEvents;
-use Sulu\Bundle\TagBundle\Event\TagMergedEvent;
 use Sulu\Bundle\TagBundle\Event\TagMergeEvent;
-use Sulu\Bundle\TagBundle\Event\TagModifiedEvent;
-use Sulu\Bundle\TagBundle\Event\TagRemovedEvent;
 use Sulu\Bundle\TagBundle\Tag\Exception\TagAlreadyExistsException;
 use Sulu\Bundle\TagBundle\Tag\Exception\TagNotFoundException;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
